removes more unused code
